+++ src/test/kotlin/DateTimeTest.kt
@@ -1,0 +1,15 @@
+package no.nav
+
+import io.kotest.matchers.nulls.shouldNotBeNull
+import no.nav.kafka.convertToOffsetDatetime
+import org.junit.Test
+
+class DateTimeTest {
+    @Test
+    fun testDateTimeParse() {
+        val dateTimeString = "2025-04-10T13:01:14+02"
+        val localDateTime = dateTimeString.convertToOffsetDatetime()
+        localDateTime.shouldNotBeNull()
+        println(localDateTime)
+    }
+}
